eu-hotel-scanner-v0.7/eu-hotel-scanner-v0.7/requirements.txt
PyYAML>=6.0
requests>=2.31
streamlit>=1.32
python-dotenv>=1.0
